requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==4.1.0
pymupdf
python-dotenv==1.1.0
langchain-community==0.3.26
faiss-cpu
langchain-google-genai
-e.

